# swagger.json
openapi: 3.0.0
info:
  title: Backgammon Game API
  description: |
    API for a real‑time multiplayer backgammon game.
    - REST endpoints for account management, matchmaking, history, shop, tournament, etc.
    - WebSocket connection for live game events (see WebSocket section).
    - Admin endpoints for managing inventory, shop, starter packs and user inventory.
    
    **Note:** Tournament game results are automatically recorded by the server; clients do not need to send `tournament.monthly.record`.
    
    **OTP (Fake):** For development, OTP endpoints return the generated code in the response. In production, the code is sent via SMS and not returned.
    
    **Debug API:** A special endpoint `/api/debug/simulate-game` is available only in development environment to simulate game results for testing progression and economy without playing a full match.
  version: 1.0.0
  contact:
    name: Support
    url: https://example.com
servers:
  - url: http://localhost:8080
    description: Development server
  - url: http://37.255.218.236:5208
    description: Production server
...
paths:
...
  /api/tournament/current/{type}:
    get:
      tags: [Tournament]
      summary: Get current tournament season
      description: Returns active season for given type (WEEKLY or MONTHLY)
      parameters:
        - name: type
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/TournamentType'
      responses:
        '200':
          description: Current season data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentSeason'
        '404':
          description: No active season found
        '500':
          description: Server error
  /api/tournament/monthly/leaderboard/{seasonId}:
    get:
      tags: [Tournament]
      summary: Get monthly tournament leaderboard
      description: Returns leaderboard for a given monthly season with tie‑breakers applied.
      parameters:
        - name: seasonId
          in: path
          required: true
          schema:
            type: integer
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 100
      responses:
        '200':
          description: Leaderboard entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TournamentLeaderboardEntry'
        '500':
          description: Server error
  /api/tournament/weekly/leaderboard/{seasonId}:
    get:
      tags: [Tournament]
      summary: Get weekly tournament leaderboard
      description: Returns leaderboard for a given weekly season. Only players with at least 5 games are included.
      parameters:
        - name: seasonId
          in: path
          required: true
          schema:
            type: integer
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 100
      responses:
        '200':
          description: Leaderboard entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TournamentLeaderboardEntry'
        '500':
          description: Server error
  /api/tournament/monthly/active:
    get:
      tags: [Tournament]
      summary: Get current user's active monthly series
      description: Returns the active series (24h window) for the authenticated user, or null if none.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Active series or null
          content:
            application/json:
              schema:
                nullable: true
                allOf:
                  - $ref: '#/components/schemas/TournamentSeries'
        '401':
          description: Not authenticated
        '500':
          description: Server error
  /api/tournament/monthly/history/{playerId}:
    get:
      tags: [Tournament]
      summary: Get monthly series history for a player
      parameters:
        - name: playerId
          in: path
          required: true
          schema:
            type: integer
        - name: seasonId
          in: query
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: List of recorded series
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TournamentSeries'
        '500':
          description: Server error
  /api/tournament/ticket/balance:
    get:
      tags: [Tournament]
      summary: Get current user's golden ticket balance (legacy)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Ticket balance
          content:
            application/json:
              schema:
                type: object
                properties:
                  balance:
                    type: integer
                    example: 5
        '401':
          description: Not authenticated
        '500':
          description: Server error
  /api/tournament/ticket/info:
    get:
      tags: [Tournament]
      summary: Get detailed ticket information
      description: Returns balance, next recharge time, max balance, and interval in seconds.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Ticket info
          content:
            application/json:
              schema:
                type: object
                properties:
                  balance:
                    type: integer
                    example: 3
                  nextRegenAt:
                    type: string
                    format: date-time
                    example: "2025-01-15T00:00:00.000Z"
                  remainingSeconds:
                    type: integer
                    example: 43200
                  regenIntervalHours:
                    type: integer
                    example: 24
                  maxBalance:
                    type: integer
                    example: 5
        '401':
          description: Not authenticated
        '500':
          description: Server error
  /api/tournament/weekly/stats:
    get:
      tags: [Tournament]
      summary: Get weekly tournament stats for current user
      description: Returns total games, Best‑20 score, best games list, and eligibility status.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Weekly stats
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalGames:
                    type: integer
                    example: 7
                  bestScore:
                    type: integer
                    example: 185
                  bestGames:
                    type: array
                    items:
                      type: object
                      properties:
                        gameId:
                          type: integer
                        result:
                          $ref: '#/components/schemas/MatchResultType'
                        totalScore:
                          type: integer
                        finishedAt:
                          type: string
                          format: date-time
                  eligibility:
                    type: object
                    properties:
                      requiredGames:
                        type: integer
                        example: 5
                      played:
                        type: integer
                        example: 7
                      remaining:
                        type: integer
                        example: 0
                      isEligible:
                        type: boolean
                        example: true
        '401':
          description: Not authenticated
        '404':
          description: No active weekly season
        '500':
          description: Server error

  # ===================== SPIN (Lucky Wheel) =====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    TournamentType:
      type: string
      enum: [WEEKLY, MONTHLY]
    TournamentStatus:
      type: string
      enum: [ACTIVE, FINALIZED]
    SeriesStatus:
      type: string
      enum: [ACTIVE, COMPLETED, CLOSED_BY_PLAYER, EXPIRED]
    TournamentSeason:
      type: object
      properties:
        id:
          type: integer
        type:
          $ref: '#/components/schemas/TournamentType'
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        status:
          $ref: '#/components/schemas/TournamentStatus'
        ticketPrice:
          type: integer
          description: Entry cost in golden tickets
        minGamesForRanking:
          type: integer
          description: Minimum games required for weekly leaderboard eligibility
        bestCount:
          type: integer
          description: Number of best results counted (weekly=20, monthly=5)
        prizePool:
          type: object
          description: JSON with prize distribution (first, top10, milestones)
          example: { first: 1000, top10: 100, milestone10: 50, milestone15: 75 }
    TournamentGame:
      type: object
      properties:
        id:
          type: integer
        seriesId:
          type: integer
        matchIndex:
          type: integer
        gameId:
          type: integer
          nullable: true
        result:
          $ref: '#/components/schemas/MatchResultType'
        baseScore:
          type: integer
        pipBonus:
          type: integer
        cleanBonus:
          type: integer
        totalScore:
          type: integer
        isForfeit:
          type: boolean
        isValid:
          type: boolean
        finishedAt:
          type: string
          format: date-time
          nullable: true
    MatchResultType:
      type: string
      enum: [normal, gammon, backgammon, timer, doublingcube, leave, forfeit, loss]
    TournamentSeries:
      type: object
      properties:
        id:
          type: integer
        playerId:
          type: integer
        seasonId:
          type: integer
        status:
          $ref: '#/components/schemas/SeriesStatus'
        startedAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
        closedAt:
          type: string
          format: date-time
          nullable: true
        totalScore:
          type: integer
        isRecorded:
          type: boolean
        gamesCount:
          type: integer
        bestSingleScore:
          type: integer
          nullable: true
        totalBackgammonWins:
          type: integer
        totalMarsWins:
          type: integer
        totalPipAdvantage:
          type: integer
        totalForfeits:
          type: integer
        achievedAt:
          type: string
          format: date-time
          nullable: true
        games:
          type: array
          items:
            $ref: '#/components/schemas/TournamentGame'
    TournamentLeaderboardEntry:
      type: object
      properties:
        id:
          type: integer
        seasonId:
          type: integer
        playerId:
          type: integer
        finalScore:
          type: integer
        bestSeriesIds:
          type: array
          items:
            type: integer
        bestSingleScore:
          type: integer
        totalBackgammonWins:
          type: integer
        totalMarsWins:
          type: integer
        avgPipAdvantage:
          type: number
        totalForfeits:
          type: integer
        seriesCount:
          type: integer
        achievedAt:
          type: string
          format: date-time
          nullable: true
        rank:
          type: integer
          nullable: true
        player:
          type: object
          properties:
            id:
              type: integer
            userName:
              type: string
            avatar:
              type: string
              nullable: true
            title:
              type: string
              nullable: true

    # ---------- USER STATS ----------
...